refactor: isolate Unix session platform mechanics
