+++ crates/cleat/src/lib.rs
@@ -4,6 +4,7 @@ pub mod cli;
 pub mod da;
 pub mod duration_parser;
 pub mod keys;
+pub mod platform;
 pub mod protocol;
 pub mod recording;
 pub mod replay;
