+++ crates/cleat/src/platform/mod.rs
@@ -1,0 +1,6 @@
+pub mod daemon;
+pub mod ipc;
+pub mod terminal;
+
+#[cfg(unix)]
+pub mod unix;
