+++ crates/cleat/src/session.rs
@@ -1,15 +1,8 @@
 use std::{
     collections::VecDeque,
-    ffi::CString,
     fs,
     io::{Read, Write},
-    net::Shutdown,
-    os::{
-        fd::{AsRawFd, BorrowedFd, IntoRawFd, RawFd},
-        unix::{fs::PermissionsExt, net::UnixStream},
-    },
     path::{Path, PathBuf},
-    process::{Command, Stdio},
     sync::{
         atomic::{AtomicBool, Ordering},
         Arc, Mutex,
@@ -18,40 +11,37 @@ use std::{
     time::{Duration, Instant},
 };
 
-use nix::{
-    errno::Errno,
-    fcntl::{fcntl, FcntlArg, OFlag},
-    poll::{poll, PollFd, PollFlags, PollTimeout},
-    pty::{forkpty, ForkptyResult},
-    sys::{
-        termios::{self, SetArg},
-        wait::{waitpid, WaitPidFlag, WaitStatus},
-    },
-    unistd::{chdir, execvp, isatty, read as nix_read, write as nix_write, Pid},
-};
-
+#[cfg(unix)]
+use crate::platform::unix::{exit_code_from_wait_status, listener_fd, poll_session_ready, stream_fd, PtyChild};
 use crate::{
     da::DeviceAttributeTracker,
+    platform::{
+        daemon::spawn_daemon_process,
+        ipc::{
+            bind_session_listener, connect_session_stream, session_socket_path as platform_session_socket_path, set_listener_nonblocking,
+            set_stream_nonblocking, set_stream_read_timeout, shutdown_stream, SessionStream,
+        },
+        terminal::{
+            attach_signal_exit_requested, current_terminal_size, poll_stdin_readable, stdout_is_tty, AttachSignalHandlers,
+            TerminalModeGuard,
+        },
+    },
     protocol::Frame,
     runtime::{RuntimeLayout, SessionMetadata},
     vt::{self, ScreenGrid, VtEngine, VtEngineKind},
 };
 
-const SOCKET_NAME: &str = "socket";
-const PID_NAME: &str = "daemon.pid";
 const FOREGROUND_NAME: &str = "foreground";
-const STRIP_ENV_VARS: &[&str] = &["SSH_TTY", "SSH_CONNECTION", "SSH_CLIENT"];
 const DEFAULT_TERMINAL_COLS: u16 = 80;
 const DEFAULT_TERMINAL_ROWS: u16 = 24;
 const DETACH_CLEANUP_SEQUENCE: &[u8] = b"\x1b[?1000l\x1b[?1002l\x1b[?1003l\x1b[?1006l\x1b[?2004l\x1b[?1004l\x1b[?1049l\x1b[<u\x1b[?25h";
 const REATTACH_CLEAR_SEQUENCE: &[u8] = b"\x1b[2J\x1b[H";
 const PTY_READ_BUFFER_SIZE: usize = 64 * 1024;
 const MAX_PENDING_CLIENT_OUTPUT_BYTES: usize = 4 * 1024 * 1024;
-static ATTACH_SIGNAL_EXIT: AtomicBool = AtomicBool::new(false);
 
 #[derive(Debug)]
 pub struct ForegroundAttach {
-    stream: Arc<Mutex<UnixStream>>,
+    stream: Arc<Mutex<SessionStream>>,
 }
 
 impl ForegroundAttach {
@@ -103,16 +93,15 @@ impl ForegroundAttach {
         });
 
         let mut stdin = std::io::stdin().lock();
-        let stdin_fd = stdin.as_raw_fd();
         let mut buf = [0u8; 4096];
         let stdin_result = loop {
-            if !alive.load(Ordering::SeqCst) || ATTACH_SIGNAL_EXIT.load(Ordering::SeqCst) {
+            if !alive.load(Ordering::SeqCst) || attach_signal_exit_requested() {
                 break Ok(());
             }
-            match poll_fd_readable(stdin_fd, 100) {
+            match poll_stdin_readable(Duration::from_millis(100)) {
                 Ok(false) => continue,
                 Ok(true) => {}
-                Err(_err) if ATTACH_SIGNAL_EXIT.load(Ordering::SeqCst) => break Ok(()),
+                Err(_err) if attach_signal_exit_requested() => break Ok(()),
                 Err(err) => break Err(err),
             }
             match stdin.read(&mut buf) {
@@ -131,10 +120,10 @@ impl ForegroundAttach {
             }
         };
 
-        let signal_exit = ATTACH_SIGNAL_EXIT.load(Ordering::SeqCst);
+        let signal_exit = attach_signal_exit_requested();
         alive.store(false, Ordering::SeqCst);
         if let Ok(stream) = self.stream.lock() {
-            let _ = stream.shutdown(Shutdown::Both);
+            shutdown_stream(&stream);
         }
         let out_result = relay_out.join().map_err(|_| "stdout relay thread panicked".to_string())?;
         let resize_result = resize_loop.join().map_err(|_| "resize thread panicked".to_string())?;
@@ -168,47 +157,6 @@ struct AttachCleanupGuard {
     target: AttachCleanupTarget,
     enabled: bool,
     emitted: bool,
-}
-
-struct AttachSignalHandlers {
-    previous: Vec<(libc::c_int, libc::sigaction)>,
-}
-
-impl AttachSignalHandlers {
-    fn install() -> Result<Self, String> {
-        ATTACH_SIGNAL_EXIT.store(false, Ordering::SeqCst);
-        let mut previous = Vec::new();
-        for signal in [libc::SIGINT, libc::SIGTERM, libc::SIGHUP] {
-            let mut old = unsafe { std::mem::zeroed::<libc::sigaction>() };
-            let mut new = unsafe { std::mem::zeroed::<libc::sigaction>() };
-            new.sa_sigaction = attach_signal_handler as *const () as usize;
-            new.sa_flags = 0;
-            unsafe {
-                libc::sigemptyset(&mut new.sa_mask);
-            }
-            let rc = unsafe { libc::sigaction(signal, &new, &mut old) };
-            if rc != 0 {
-                return Err(format!("install signal handler {signal}: {}", std::io::Error::last_os_error()));
-            }
-            previous.push((signal, old));
-        }
-        Ok(Self { previous })
-    }
-}
-
-impl Drop for AttachSignalHandlers {
-    fn drop(&mut self) {
-        ATTACH_SIGNAL_EXIT.store(false, Ordering::SeqCst);
-        for (signal, action) in self.previous.drain(..).rev() {
-            unsafe {
-                libc::sigaction(signal, &action, std::ptr::null_mut());
-            }
-        }
-    }
-}
-
-extern "C" fn attach_signal_handler(_signal: libc::c_int) {
-    ATTACH_SIGNAL_EXIT.store(true, Ordering::SeqCst);
 }
 
 impl AttachCleanupGuard {
@@ -262,46 +210,6 @@ fn write_detach_cleanup<W: Write>(writer: &mut W) -> Result<(), String> {
     writer.flush().map_err(|err| format!("flush detach cleanup: {err}"))
 }
 
-fn stdout_is_tty() -> Result<bool, String> {
-    let fd = std::io::stdout().as_raw_fd();
-    // SAFETY: stdout remains open for the duration of this check; we only borrow its fd.
-    let borrowed_fd = unsafe { BorrowedFd::borrow_raw(fd) };
-    isatty(borrowed_fd).map_err(|err| format!("detect terminal stdout: {err}"))
-}
-
-struct TerminalModeGuard {
-    fd: RawFd,
-    original: Option<termios::Termios>,
-}
-
-impl TerminalModeGuard {
-    fn activate() -> Result<Self, String> {
-        let fd = std::io::stdin().as_raw_fd();
-        // SAFETY: stdin remains open for the lifetime of the guard; we only borrow its fd.
-        let borrowed_fd = unsafe { BorrowedFd::borrow_raw(fd) };
-        if !isatty(borrowed_fd).map_err(|err| format!("detect terminal stdin: {err}"))? {
-            return Ok(Self { fd, original: None });
-        }
-
-        let original = termios::tcgetattr(borrowed_fd).map_err(|err| format!("read terminal attrs: {err}"))?;
-        let mut raw = original.clone();
-        termios::cfmakeraw(&mut raw);
-        termios::tcsetattr(borrowed_fd, SetArg::TCSAFLUSH, &raw).map_err(|err| format!("set terminal raw mode: {err}"))?;
-
-        Ok(Self { fd, original: Some(original) })
-    }
-}
-
-impl Drop for TerminalModeGuard {
-    fn drop(&mut self) {
-        if let Some(original) = self.original.as_ref() {
-            // SAFETY: stdin remains open for the lifetime of the guard; we only borrow its fd.
-            let borrowed_fd = unsafe { BorrowedFd::borrow_raw(self.fd) };
-            let _ = termios::tcsetattr(borrowed_fd, SetArg::TCSAFLUSH, original);
-        }
-    }
-}
-
 pub fn ensure_session_started(
     layout: &RuntimeLayout,
     id: Option<String>,
@@ -338,7 +246,7 @@ pub fn attach_foreground(layout: &RuntimeLayout, id: &str) -> Result<ForegroundA
     let socket_path = session_socket_path(layout.root(), id);
     let deadline = Instant::now() + Duration::from_millis(250);
     loop {
-        let mut stream = UnixStream::connect(&socket_path).map_err(|err| format!("connect {}: {err}", socket_path.display()))?;
+        let mut stream = connect_session_stream(&socket_path)?;
         let (cols, rows) = current_terminal_size();
         Frame::AttachInit { cols, rows, capabilities: attach_init_capabilities() }
             .write(&mut stream)
@@ -356,11 +264,11 @@ pub fn attach_foreground(layout: &RuntimeLayout, id: &str) -> Result<ForegroundA
 }
 
 pub fn session_socket_path(root: &Path, id: &str) -> PathBuf {
-    root.join(id).join(SOCKET_NAME)
+    platform_session_socket_path(root, id)
 }
 
 pub fn daemon_pid_path(root: &Path, id: &str) -> PathBuf {
-    root.join(id).join(PID_NAME)
+    crate::platform::daemon::daemon_pid_path(root, id)
 }
 
 pub fn foreground_path(root: &Path, id: &str) -> PathBuf {
@@ -449,14 +357,14 @@ fn apply_attach_state(
 }
 
 struct PendingWait {
-    stream: UnixStream,
+    stream: SessionStream,
     conditions: Vec<crate::protocol::WaitCondition>,
     timeout_ms: u64,
     registered_at: Instant,
 }
 
 struct PendingExpect {
-    stream: UnixStream,
+    stream: SessionStream,
     text: String,
     since_offset: u64,
     last_checked_file_size: u64,
@@ -474,14 +382,12 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
         let _ = fs::remove_file(&socket_path);
     }
 
-    let listener =
-        std::os::unix::net::UnixListener::bind(&socket_path).map_err(|err| format!("bind socket {}: {err}", socket_path.display()))?;
-    listener.set_nonblocking(true).map_err(|err| format!("set listener nonblocking: {err}"))?;
+    let listener = bind_session_listener(&socket_path)?;
+    set_listener_nonblocking(&listener, true)?;
     fs::write(daemon_pid_path(root, id), std::process::id().to_string()).map_err(|err| format!("write daemon pid: {err}"))?;
 
-    let pty_child = spawn_pty_child(session)?;
-    let pty_fd = pty_child.master_fd;
-    set_nonblocking(pty_fd)?;
+    let pty_child = PtyChild::spawn(session)?;
+    pty_child.set_nonblocking()?;
     let mut vt_engine = default_vt_engine(session.vt_engine)?;
     // The DA tracker is the only DA source for the passthrough engine.
     // The ghostty engine answers DA itself via its DeviceAttributes callback,
@@ -514,11 +420,11 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
     let mut pending_expects: Vec<PendingExpect> = Vec::new();
     let mut last_pty_output_at: Option<Instant> = None;
     loop {
-        let poll_result = poll_ready(
-            listener.as_raw_fd(),
-            active_client.as_ref().map(|client| client.stream.as_raw_fd()),
+        let poll_result = poll_session_ready(
+            listener_fd(&listener),
+            active_client.as_ref().map(|client| stream_fd(&client.stream)),
             active_client.as_ref().map(|client| !client.pending_output.is_empty()).unwrap_or(false),
-            pty_fd,
+            pty_child.master_fd(),
             100,
         )?;
 
@@ -527,15 +433,15 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
                 Ok((mut stream, _)) => {
                     // Accepted sockets inherit nonblocking mode from the listener on macOS/BSD.
                     // Reset to blocking so the initial frame read works correctly.
-                    stream.set_nonblocking(false).map_err(|err| format!("set accepted stream blocking: {err}"))?;
-                    let _ = stream.set_read_timeout(Some(Duration::from_millis(100)));
+                    set_stream_nonblocking(&stream, false).map_err(|err| format!("set accepted stream blocking: {err}"))?;
+                    let _ = set_stream_read_timeout(&stream, Some(Duration::from_millis(100)));
                     match Frame::read(&mut stream) {
                         Ok(Frame::AttachInit { cols, rows, capabilities }) => {
                             if active_client.is_none() {
-                                resize_pty(pty_fd, cols, rows)?;
+                                pty_child.resize(cols, rows)?;
                                 let replay = apply_attach_state(vt_engine.as_mut(), cols, rows, &capabilities)?;
                                 Frame::Ack.write(&mut stream).map_err(|err| format!("write attach ack: {err}"))?;
-                                stream.set_nonblocking(true).map_err(|err| format!("set client nonblocking: {err}"))?;
+                                set_stream_nonblocking(&stream, true).map_err(|err| format!("set client nonblocking: {err}"))?;
                                 let mut client = ActiveClient::new(stream);
                                 if let Some(payload) = replay {
                                     if !payload.is_empty() {
@@ -574,7 +480,7 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
                             if let Some(ref mut rec) = recorder {
                                 rec.input(&bytes, epoch.elapsed());
                             }
-                            if let Err(err) = write_fd_all(pty_fd, &bytes) {
+                            if let Err(err) = pty_child.write_all(&bytes) {
                                 let _ = Frame::Error(err).write(&mut stream);
                             }
                         }
@@ -585,7 +491,7 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
                                 let offset = rec.bytes_written();
                                 markers.insert(marker_name, offset);
                                 rec.input(&bytes, epoch.elapsed());
-                                if let Err(err) = write_fd_all(pty_fd, &bytes) {
+                                if let Err(err) = pty_child.write_all(&bytes) {
                                     let _ = Frame::Error(err).write(&mut stream);
                                 } else {
                                     let _ = Frame::MarkResult { offset }.write(&mut stream);
@@ -605,7 +511,7 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
                                 }
                             }
                         }
-                        Ok(Frame::Signal { signal, target }) => match dispatch_signal(&pty_child, signal, target) {
+                        Ok(Frame::Signal { signal, target }) => match pty_child.dispatch_signal(signal, target) {
                             Ok(()) => {
                                 if let Some(ref mut rec) = recorder {
                                     let target_str = match target {
@@ -735,7 +641,7 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
                             }
 
                             // Register for async evaluation
-                            if let Err(err) = stream.set_nonblocking(true) {
+                            if let Err(err) = set_stream_nonblocking(&stream, true) {
                                 let _ = Frame::Error(format!("set nonblocking: {err}")).write(&mut stream);
                                 break 'wait;
                             }
@@ -761,7 +667,7 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
                                     }
                                 }
                             }
-                            if let Err(err) = stream.set_nonblocking(true) {
+                            if let Err(err) = set_stream_nonblocking(&stream, true) {
                                 let _ = Frame::Error(format!("set nonblocking: {err}")).write(&mut stream);
                                 break 'expect;
                             }
@@ -819,13 +725,13 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
                         if let Some(ref mut rec) = recorder {
                             rec.input(&bytes, epoch.elapsed());
                         }
-                        write_fd_all(pty_fd, &bytes)?;
+                        pty_child.write_all(&bytes)?;
                     }
                     Frame::Resize { cols, rows } => {
                         if let Some(ref mut rec) = recorder {
                             rec.event(crate::asciicast::EventCode::Resize, &format!("{}x{}", cols, rows), epoch.elapsed());
                         }
-                        resize_pty(pty_fd, cols, rows)?;
+                        pty_child.resize(cols, rows)?;
                         vt_engine.resize(cols, rows)?;
                     }
                     _ => {}
@@ -858,7 +764,7 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
         if poll_result.pty_readable {
             loop {
                 let mut buf = [0u8; PTY_READ_BUFFER_SIZE];
-                match read_fd(pty_fd, &mut buf) {
+                match pty_child.read_output(&mut buf) {
                     Ok(0) => break,
                     Ok(n) => {
                         last_pty_output_at = Some(Instant::now());
@@ -883,11 +789,11 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
                         if active_client.is_none() {
                             if let Some(ref mut tracker) = detached_da {
                                 for reply in tracker.push(&buf[..n]) {
-                                    write_fd_all(pty_fd, &reply)?;
+                                    pty_child.write_all(&reply)?;
                                 }
                             }
                             if !engine_reply.is_empty() {
-                                write_fd_all(pty_fd, &engine_reply)?;
+                                pty_child.write_all(&engine_reply)?;
                             }
                         }
                         if let Some(client) = active_client.as_mut() {
@@ -988,7 +894,7 @@ pub fn run_session_daemon(root: &Path, session: &SessionMetadata) -> Result<(), 
             });
         }
 
-        if let Some(status) = child_exited(pty_child.pid)? {
+        if let Some(status) = pty_child.exited()? {
             for mut wait in pending_waits.drain(..) {
                 let elapsed_ms = wait.registered_at.elapsed().as_millis() as u64;
                 let _ = Frame::WaitResult { status: crate::protocol::WaitStatus::SessionGone, elapsed_ms }.write(&mut wait.stream);
@@ -1023,41 +929,6 @@ pub fn run_session_daemon(_root: &Path, _session: &SessionMetadata) -> Result<()
     Err("session daemon is only supported on unix".into())
 }
 
-/// Resolve the current working directory for a given process ID.
-/// Returns `None` if the pid is invalid or the cwd cannot be determined.
-#[cfg(target_os = "linux")]
-fn resolve_cwd(pid: u32) -> Option<std::path::PathBuf> {
-    std::fs::read_link(format!("/proc/{pid}/cwd")).ok()
-}
-
-#[cfg(target_os = "macos")]
-fn resolve_cwd(pid: u32) -> Option<std::path::PathBuf> {
-    use std::mem;
-    // SAFETY: we zero-initialize the struct and pass valid arguments to proc_pidinfo.
-    // proc_pidinfo is a well-known macOS API for querying process info.
-    unsafe {
-        let mut vnode_info: libc::proc_vnodepathinfo = mem::zeroed();
-        let size = mem::size_of::<libc::proc_vnodepathinfo>() as libc::c_int;
-        let ret =
-            libc::proc_pidinfo(pid as libc::c_int, libc::PROC_PIDVNODEPATHINFO, 0, &mut vnode_info as *mut _ as *mut libc::c_void, size);
-        if ret <= 0 {
-            return None;
-        }
-        let cstr = std::ffi::CStr::from_ptr(vnode_info.pvi_cdir.vip_path.as_ptr() as *const libc::c_char);
-        let path = std::path::PathBuf::from(cstr.to_string_lossy().into_owned());
-        if path.as_os_str().is_empty() {
-            None
-        } else {
-            Some(path)
-        }
-    }
-}
-
-#[cfg(not(any(target_os = "linux", target_os = "macos")))]
-fn resolve_cwd(_pid: u32) -> Option<std::path::PathBuf> {
-    None
-}
-
 fn build_inspect_result(
     session: &SessionMetadata,
     vt_engine: &dyn VtEngine,
@@ -1067,9 +938,7 @@ fn build_inspect_result(
     markers: &std::collections::HashMap<String, u64>,
 ) -> crate::protocol::InspectResult {
     let (cols, rows) = vt_engine.size();
-    // SAFETY: pty_child.master_fd is a valid PTY master fd owned by this process.
-    let foreground_pgid =
-        nix::unistd::tcgetpgrp(unsafe { std::os::fd::BorrowedFd::borrow_raw(pty_child.master_fd) }).ok().map(|pid| pid.as_raw() as u32);
+    let foreground_pgid = pty_child.foreground_pgid();
 
     crate::protocol::InspectResult {
         session: crate::protocol::SessionInspect {
@@ -1083,11 +952,10 @@ fn build_inspect_result(
         },
         terminal: crate::protocol::TerminalInspect { rows, cols },
         process: crate::protocol::ProcessInspect {
-            leader_pid: pty_child.pid.as_raw() as u32,
+            leader_pid: pty_child.leader_pid(),
             foreground_pgid,
-            leader_cwd: resolve_cwd(pty_child.pid.as_raw() as u32),
-            // PGID equals the group leader's PID, so resolve_cwd works here.
-            foreground_cwd: foreground_pgid.and_then(resolve_cwd),
+            leader_cwd: pty_child.leader_cwd(),
+            foreground_cwd: pty_child.foreground_cwd(),
         },
         attachments: if active_client.is_some() {
             vec![crate::protocol::AttachmentInspect { role: "controller".to_string() }]
@@ -1102,97 +970,13 @@ fn build_inspect_result(
     }
 }
 
-fn dispatch_signal(pty_child: &PtyChild, signal: i32, target: crate::protocol::SignalTarget) -> Result<(), String> {
-    use nix::sys::signal::{killpg, Signal};
-
-    let signal = Signal::try_from(signal).map_err(|err| format!("invalid signal number: {err}"))?;
-
-    match target {
-        crate::protocol::SignalTarget::Foreground => {
-            // SAFETY: pty_child.master_fd is a valid PTY master fd owned by this process.
-            let fg_pgid = nix::unistd::tcgetpgrp(unsafe { std::os::fd::BorrowedFd::borrow_raw(pty_child.master_fd) })
-                .map_err(|err| format!("tcgetpgrp: {err}"))?;
-            killpg(fg_pgid, signal).map_err(|err| format!("killpg: {err}"))
-        }
-        crate::protocol::SignalTarget::Leader => nix::sys::signal::kill(pty_child.pid, signal).map_err(|err| format!("kill: {err}")),
-        crate::protocol::SignalTarget::Tree => Err("tree signal target is not yet implemented".to_string()),
-    }
-}
-
-fn spawn_daemon_process(root: &Path, session: &SessionMetadata) -> Result<(), String> {
-    let exe = resolve_cleat_executable()?;
-    let mut command = Command::new(exe);
-    command.arg("--runtime-root").arg(root).arg("serve").arg("--id").arg(&session.id).arg("--vt").arg(session.vt_engine.as_str());
-    if let Some(cmd) = &session.cmd {
-        command.arg("--cmd").arg(cmd);
-    }
-    if let Some(cwd) = &session.cwd {
-        command.arg("--cwd").arg(cwd);
-    }
-    if session.record {
-        command.arg("--record");
-    }
-    command.stdin(Stdio::null()).stdout(Stdio::null()).stderr(Stdio::null());
-    let child = command.spawn().map_err(|err| format!("spawn session daemon for {}: {err}", session.id))?;
-    fs::write(daemon_pid_path(root, &session.id), child.id().to_string()).map_err(|err| format!("write daemon pid: {err}"))?;
-    Ok(())
-}
-
-fn resolve_cleat_executable() -> Result<PathBuf, String> {
-    if let Some(path) = std::env::var_os("CARGO_BIN_EXE_cleat").map(PathBuf::from) {
-        return Ok(path);
-    }
-
-    let sibling = current_exe_sibling("cleat");
-    let path_var = std::env::var_os("PATH").unwrap_or_default();
-
-    resolve_cleat_with_sibling(sibling.as_deref(), &path_var)
-}
-
-fn resolve_cleat_with_sibling(sibling: Option<&Path>, path_var: &std::ffi::OsStr) -> Result<PathBuf, String> {
-    // Prefer sibling of current executable — strongest "same version" signal.
-    if let Some(path) = sibling {
-        if is_executable_file(path) {
-            return Ok(path.to_path_buf());
-        }
-    }
-
-    // Fall back to PATH search.
-    for dir in std::env::split_paths(path_var) {
-        let candidate = dir.join("cleat");
-        if is_executable_file(&candidate) {
-            return Ok(candidate);
-        }
-    }
-
-    Err("unable to locate cleat executable on PATH or next to current binary".into())
-}
-
-fn current_exe_sibling(name: &str) -> Option<PathBuf> {
-    let current_exe = std::env::current_exe().ok()?;
-    let current_dir = current_exe.parent()?;
-    let candidates = [current_dir.join(name), current_dir.parent().map(|parent| parent.join(name))?];
-    candidates.into_iter().find(|candidate| is_executable_file(candidate))
-}
-
-fn is_executable_file(path: &Path) -> bool {
-    path.is_file() && fs::metadata(path).map(|metadata| metadata.permissions().mode() & 0o111 != 0).unwrap_or(false)
-}
-
-struct PollResult {
-    listener_readable: bool,
-    client_readable: bool,
-    client_writable: bool,
-    pty_readable: bool,
-}
-
 struct ActiveClient {
-    stream: UnixStream,
+    stream: SessionStream,
     pending_output: Vec<u8>,
 }
 
 impl ActiveClient {
-    fn new(stream: UnixStream) -> Self {
+    fn new(stream: SessionStream) -> Self {
         Self { stream, pending_output: Vec::new() }
     }
 
@@ -1222,59 +1006,6 @@ impl ActiveClient {
     }
 }
 
-fn poll_fd_readable(fd: RawFd, timeout_ms: i32) -> Result<bool, String> {
-    // SAFETY: the fd remains open for the duration of the poll call; we only borrow it temporarily.
-    let borrowed = unsafe { BorrowedFd::borrow_raw(fd) };
-    let mut fds = [PollFd::new(borrowed, PollFlags::POLLIN)];
-    match poll(&mut fds, PollTimeout::try_from(timeout_ms).map_err(|err| format!("invalid poll timeout: {err}"))?) {
-        Ok(_) => {}
-        Err(Errno::EINTR) => return Ok(false),
-        Err(err) => return Err(format!("poll readable fd: {err}")),
-    }
-    Ok(has_pollin(&fds[0]))
-}
-
-struct PtyChild {
-    master_fd: RawFd,
-    pid: Pid,
-}
-
-fn poll_ready(
-    listener_fd: RawFd,
-    client_fd: Option<RawFd>,
-    client_needs_write: bool,
-    pty_fd: RawFd,
-    timeout_ms: i32,
-) -> Result<PollResult, String> {
-    // SAFETY: the fds are owned by this process and remain open for the duration of the poll call.
-    let listener_borrowed = unsafe { BorrowedFd::borrow_raw(listener_fd) };
-    // SAFETY: the fd is owned by this process and remains open for the duration of the poll call.
-    let pty_borrowed = unsafe { BorrowedFd::borrow_raw(pty_fd) };
-    let mut fds = vec![PollFd::new(listener_borrowed, PollFlags::POLLIN), PollFd::new(pty_borrowed, PollFlags::POLLIN)];
-    let client_index = if let Some(fd) = client_fd {
-        // SAFETY: the client fd is owned by this process and remains open for the duration of the poll call.
-        let client_borrowed = unsafe { BorrowedFd::borrow_raw(fd) };
-        let mut flags = PollFlags::POLLIN;
-        if client_needs_write {
-            flags |= PollFlags::POLLOUT;
-        }
-        fds.push(PollFd::new(client_borrowed, flags));
-        Some(fds.len() - 1)
-    } else {
-        None
-    };
-
-    poll(&mut fds, PollTimeout::try_from(timeout_ms).map_err(|err| format!("invalid poll timeout: {err}"))?)
-        .map_err(|err| format!("poll daemon fds: {err}"))?;
-
-    Ok(PollResult {
-        listener_readable: has_pollin(&fds[0]),
-        pty_readable: has_pollin(&fds[1]),
-        client_readable: client_index.map(|index| has_pollin(&fds[index])).unwrap_or(false),
-        client_writable: client_index.map(|index| has_pollout(&fds[index])).unwrap_or(false),
-    })
-}
-
 fn wait_for_socket(path: &Path) -> Result<(), String> {
     let deadline = Instant::now() + Duration::from_secs(5);
     while Instant::now() < deadline {
@@ -1286,154 +1017,14 @@ fn wait_for_socket(path: &Path) -> Result<(), String> {
     Err(format!("timed out waiting for socket {}", path.display()))
 }
 
-fn has_pollin(fd: &PollFd<'_>) -> bool {
-    fd.revents().map(|flags| flags.contains(PollFlags::POLLIN)).unwrap_or(false)
-}
-
-fn has_pollout(fd: &PollFd<'_>) -> bool {
-    fd.revents().map(|flags| flags.contains(PollFlags::POLLOUT)).unwrap_or(false)
-}
-
-fn current_terminal_size() -> (u16, u16) {
-    #[cfg(unix)]
-    {
-        let fd = std::io::stdout().as_raw_fd();
-        let mut winsize = libc::winsize { ws_row: 0, ws_col: 0, ws_xpixel: 0, ws_ypixel: 0 };
-        let rc = unsafe { libc::ioctl(fd, libc::TIOCGWINSZ, &mut winsize) };
-        if rc == 0 && winsize.ws_col > 0 && winsize.ws_row > 0 {
-            return (winsize.ws_col, winsize.ws_row);
-        }
-    }
-    let cols = std::env::var("COLUMNS").ok().and_then(|value| value.parse::<u16>().ok()).unwrap_or(80);
-    let rows = std::env::var("LINES").ok().and_then(|value| value.parse::<u16>().ok()).unwrap_or(24);
-    (cols, rows)
-}
-
-#[cfg(unix)]
-fn spawn_pty_child(session: &SessionMetadata) -> Result<PtyChild, String> {
-    // SAFETY: `forkpty` creates a child attached to a new PTY; parent receives the owned master fd.
-    let result = unsafe { forkpty(None, None) }.map_err(|err| format!("forkpty failed: {err}"))?;
-    match result {
-        ForkptyResult::Parent { master, child } => Ok(PtyChild { master_fd: master.into_raw_fd(), pid: child }),
-        ForkptyResult::Child => {
-            if let Some(cwd) = &session.cwd {
-                let _ = chdir(cwd);
-            }
-            for key in STRIP_ENV_VARS {
-                // SAFETY: child process is single-threaded here, before exec, so environment mutation is safe.
-                unsafe {
-                    std::env::remove_var(key);
-                }
-            }
-            let shell = std::env::var("SHELL").unwrap_or_else(|_| "/bin/sh".into());
-            let shell_c = CString::new(shell.clone()).map_err(|_| "shell contains interior nul".to_string())?;
-            let mut args = vec![shell_c.clone()];
-            if let Some(cmd) = &session.cmd {
-                args.push(CString::new("-lc").map_err(|_| "invalid -lc".to_string())?);
-                args.push(CString::new(cmd.as_str()).map_err(|_| "cmd contains interior nul".to_string())?);
-            }
-            let _ = execvp(&shell_c, &args);
-            std::process::exit(127);
-        }
-    }
-}
-
-#[cfg(unix)]
-fn set_nonblocking(fd: RawFd) -> Result<(), String> {
-    // SAFETY: the fd is owned by this process and remains open for the duration of these fcntl calls.
-    let borrowed = unsafe { BorrowedFd::borrow_raw(fd) };
-    let flags = fcntl(borrowed, FcntlArg::F_GETFL).map_err(|err| format!("fcntl F_GETFL failed: {err}"))?;
-    let mut oflags = OFlag::from_bits_retain(flags);
-    oflags.insert(OFlag::O_NONBLOCK);
-    fcntl(borrowed, FcntlArg::F_SETFL(oflags)).map_err(|err| format!("fcntl F_SETFL failed: {err}"))?;
-    Ok(())
-}
-
-#[cfg(unix)]
-fn read_fd(fd: RawFd, buf: &mut [u8]) -> Result<usize, std::io::Error> {
-    // SAFETY: the fd is owned by this process and remains open for the duration of the read call.
-    let borrowed = unsafe { BorrowedFd::borrow_raw(fd) };
-    nix_read(borrowed, buf).map_err(std::io::Error::from)
-}
-
-#[cfg(unix)]
-fn write_fd_all(fd: RawFd, mut bytes: &[u8]) -> Result<(), String> {
-    while !bytes.is_empty() {
-        // SAFETY: the fd is owned by this process and remains open for the duration of the write call.
-        let borrowed = unsafe { BorrowedFd::borrow_raw(fd) };
-        match nix_write(borrowed, bytes) {
-            Ok(written) => bytes = &bytes[written..],
-            Err(err) => {
-                let err = std::io::Error::from(err);
-                if err.kind() == std::io::ErrorKind::WouldBlock {
-                    wait_for_writable(fd)?;
-                    continue;
-                }
-                return Err(format!("write pty input: {err}"));
-            }
-        }
-    }
-    Ok(())
-}
-
-#[cfg(unix)]
-fn wait_for_writable(fd: RawFd) -> Result<(), String> {
-    // SAFETY: the fd is owned by this process and remains open for the duration of the poll call.
-    let borrowed = unsafe { BorrowedFd::borrow_raw(fd) };
-    let mut fds = [PollFd::new(borrowed, PollFlags::POLLOUT)];
-    poll(&mut fds, PollTimeout::NONE).map_err(|err| format!("poll writable pty fd: {err}"))?;
-    Ok(())
-}
-
-#[cfg(unix)]
-fn resize_pty(fd: RawFd, cols: u16, rows: u16) -> Result<(), String> {
-    let winsize = libc::winsize { ws_row: rows, ws_col: cols, ws_xpixel: 0, ws_ypixel: 0 };
-    // SAFETY: ioctl updates the window size for a valid PTY master fd using a properly initialized winsize.
-    let rc = unsafe { libc::ioctl(fd, libc::TIOCSWINSZ, &winsize) };
-    if rc == 0 {
-        Ok(())
-    } else {
-        Err(format!("resize pty: {}", std::io::Error::last_os_error()))
-    }
-}
-
-#[cfg(unix)]
-fn child_exited(child_pid: Pid) -> Result<Option<WaitStatus>, String> {
-    match waitpid(child_pid, Some(WaitPidFlag::WNOHANG)) {
-        Ok(WaitStatus::StillAlive) => Ok(None),
-        Ok(status) => Ok(Some(status)),
-        Err(nix::errno::Errno::ECHILD) => Ok(None),
-        Err(err) => Err(format!("waitpid failed: {err}")),
-    }
-}
-
-fn exit_code_from_wait_status(status: &WaitStatus) -> i32 {
-    match status {
-        WaitStatus::Exited(_, code) => *code,
-        WaitStatus::Signaled(_, sig, _) => 128 + *sig as i32,
-        _ => 1,
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use std::{
-        fs,
-        sync::{Arc, Mutex, OnceLock},
-    };
-
-    use nix::{sys::wait::WaitStatus, unistd::Pid};
+    use std::sync::{Arc, Mutex};
 
     use super::{
-        apply_attach_state, attach_init_capabilities, default_vt_engine, exit_code_from_wait_status, is_executable_file, record_pty_output,
-        resolve_cleat_executable, AttachCleanupGuard, TestReplayProbeVtEngine,
+        apply_attach_state, attach_init_capabilities, default_vt_engine, record_pty_output, AttachCleanupGuard, TestReplayProbeVtEngine,
     };
     use crate::vt::{self, VtEngine};
-
-    fn env_lock() -> &'static Mutex<()> {
-        static LOCK: OnceLock<Mutex<()>> = OnceLock::new();
-        LOCK.get_or_init(|| Mutex::new(()))
-    }
 
     #[test]
     fn cleanup_guard_writes_on_drop() {
@@ -1544,102 +1135,5 @@ mod tests {
 
         assert_eq!(engine.size(), (100, 30));
         assert_eq!(replay, None);
-    }
-
-    #[test]
-    fn resolve_cleat_executable_prefers_cargo_bin_env() {
-        let _lock = env_lock().lock().unwrap_or_else(|e| e.into_inner());
-        let temp = tempfile::tempdir().expect("tempdir");
-        let cleat = temp.path().join("cleat");
-        fs::write(&cleat, b"#!/bin/sh\n").expect("write fake cleat");
-        let original = std::env::var_os("CARGO_BIN_EXE_cleat");
-        std::env::set_var("CARGO_BIN_EXE_cleat", &cleat);
-
-        let resolved = resolve_cleat_executable().expect("resolve cleat");
-
-        match original {
-            Some(value) => std::env::set_var("CARGO_BIN_EXE_cleat", value),
-            None => std::env::remove_var("CARGO_BIN_EXE_cleat"),
-        }
-        assert_eq!(resolved, cleat);
-    }
-
-    #[test]
-    fn resolve_cleat_executable_falls_back_to_path() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let temp = tempfile::tempdir().expect("tempdir");
-        let bin_dir = temp.path().join("bin");
-        fs::create_dir_all(&bin_dir).expect("create bin dir");
-        let cleat = bin_dir.join("cleat");
-        fs::write(&cleat, b"#!/bin/sh\n").expect("write fake cleat");
-        let mut perms = fs::metadata(&cleat).expect("metadata").permissions();
-        perms.set_mode(0o755);
-        fs::set_permissions(&cleat, perms).expect("set executable");
-
-        let resolved = super::resolve_cleat_with_sibling(None, std::ffi::OsStr::new(bin_dir.to_str().unwrap())).expect("resolve from path");
-
-        assert_eq!(resolved, cleat);
-        assert!(is_executable_file(&cleat));
-    }
-
-    #[test]
-    fn exit_code_from_normal_exit() {
-        let status = WaitStatus::Exited(Pid::from_raw(1), 42);
-        assert_eq!(exit_code_from_wait_status(&status), 42);
-    }
-
-    #[test]
-    fn exit_code_from_zero_exit() {
-        let status = WaitStatus::Exited(Pid::from_raw(1), 0);
-        assert_eq!(exit_code_from_wait_status(&status), 0);
-    }
-
-    #[test]
-    fn exit_code_from_signal_is_128_plus_signal() {
-        let status = WaitStatus::Signaled(Pid::from_raw(1), nix::sys::signal::Signal::SIGTERM, false);
-        assert_eq!(exit_code_from_wait_status(&status), 128 + libc::SIGTERM);
-    }
-
-    #[test]
-    fn resolve_cleat_exe_prefers_sibling_over_path() {
-        use std::{fs, os::unix::fs::PermissionsExt};
-
-        let temp = tempfile::tempdir().expect("tempdir");
-
-        // Create a "sibling" cleat binary
-        let sibling_dir = temp.path().join("sibling");
-        fs::create_dir_all(&sibling_dir).expect("create sibling dir");
-        let sibling_exe = sibling_dir.join("cleat");
-        fs::write(&sibling_exe, "#!/bin/sh\n").expect("write sibling");
-        fs::set_permissions(&sibling_exe, fs::Permissions::from_mode(0o755)).expect("chmod sibling");
-
-        // Create a "PATH" cleat binary
-        let path_dir = temp.path().join("path-bin");
-        fs::create_dir_all(&path_dir).expect("create path dir");
-        let path_exe = path_dir.join("cleat");
-        fs::write(&path_exe, "#!/bin/sh\n").expect("write path");
-        fs::set_permissions(&path_exe, fs::Permissions::from_mode(0o755)).expect("chmod path");
-
-        let result = super::resolve_cleat_with_sibling(Some(&sibling_exe), std::ffi::OsStr::new(path_dir.to_str().unwrap()));
-
-        assert_eq!(result.unwrap(), sibling_exe);
-    }
-
-    #[test]
-    fn resolve_cleat_exe_falls_back_to_path_when_no_sibling() {
-        use std::{fs, os::unix::fs::PermissionsExt};
-
-        let temp = tempfile::tempdir().expect("tempdir");
-
-        let path_dir = temp.path().join("path-bin");
-        fs::create_dir_all(&path_dir).expect("create path dir");
-        let path_exe = path_dir.join("cleat");
-        fs::write(&path_exe, "#!/bin/sh\n").expect("write path");
-        fs::set_permissions(&path_exe, fs::Permissions::from_mode(0o755)).expect("chmod path");
-
-        let result = super::resolve_cleat_with_sibling(None, std::ffi::OsStr::new(path_dir.to_str().unwrap()));
-
-        assert_eq!(result.unwrap(), path_exe);
     }
 }
